Fix bug dropping first word of every lesson
This is because I recently "fixed" the line that was failing to filter
out empty lines from the lesson. This started filtering out empty line
reserved for subtitles so the subsequent loop wasn't starting at the
correct index. Instead of adjusting the loop, we're removing the filter
so that it will work when there is or isn't a subtitle present.