Make Finished component responsive to handle zoom
From Plover discord:

verbalshadow 14/12/2018

@Di ⌨Typey Type  would you consider moving the option to revise the
lesson to the same location as next and retry. The inner frame is not
responsive and it's a hassle to view when z oomed or in smaller windows.

I zoom because my screen is about 3-4 ft from where i'm sitting with my
keyboard.

plus it is not fullscreen so i can see suggestions and layout