Initial height finished lesson canvas
This stops limiting the finished component height to 320px so you can
see all the possible stroke improvements easily on scroll without
needing to scroll an inner frame. This means you can more easily see the
“revise these words” link, scan the list more easily, and avoid nested
scroll areas.

One drawback of this approach is that it's much longer to reach the user
settings after completing a lesson, if that's your preferred workflow.