Update heading focus styles to simple underline
The parts of this using :not(:focus-visible) are mostly unsupported at
this time.

frag on Discord reported "I preferred the standard blue outline, I
dislike the design of it. It's ugly and distracting in my opinion, and
makes me want to constantly deselect everything that has it to get rid
of it."