Fix recommended practice your words word count bug
When you do recommended Practice lesson, "Your words", lesson sets
limitNumberOfWords to 300, then resets that to 0 via
setupRevisionLesson, which is incorrect. This approach fixes it by
checking if it's a recommended lesson and setting the revision lesson to
300. Downsides include inconsistency in variables set between
setupRevisionLesson and params.js, and setting 'practice all your words'
to have no word limit is probably a poor experience.