Fix flickering underline in typed text input area
In Chrome (not Safari or Firefox), the text in the typed text input
textarea has an underline that flickers when you have typed at least 1
character. It flicks between underlining all the typed characters and
sometimes more, as often as the cursor caret blinks. For some reason,
removing `text-underline-position: under;` seems to solve the bug,
though I can find no sensible reason for this.