Change sort and range limit to sort first
Instead of limiting the lesson range to start from word X and limit to Y
words, then sort remaining words, we'll sort the lesson first then take
the selection of words from X to X+Y for all lessons, not just
recommended lessons.

This lets us revise the newest words we've seen, not just the newest
from a subset of the oldest lessons, which is what happens if we limit
the words then sort.

The trade-off is that we can no longer take a selection, such as the top
200–300 words, then sort to randomise that selection. It will randomise
the entire lesson, then limit to the random words from 200–300.