Add extra conditional for when dict is up to date
This handles the case where we don't need the Plover dictionary for the
current feature but we've already loaded it anyway, and it's not showing
in the local config but it otherwise matches. Therefore, we don't need
to reload the globalLookupDictionary.

DictionariesIndex -> Lookup -> Lessons -> Introduction should load
Plover dict on DictionariesIndex and NOT reload it on Introduction