Remove "there are no words to write" announcement
It is announced at the wrong time. The intention was when the user
changes their "start from word" setting to, say, 100000, are the lesson
isn't that long, it should have announced the change from valid material
to "There are no words to write". Instead, after clicking "Start from
word 1", it would then announce the message, even though the button
click had fixed the issue.

This commit removes the announcement and relies on the user to manually
navigate to the material to discover there are no words to write and can
then click the "Start from word 1" button to fix that.