Remove updateFlashcardsRecommendation from Study
Previously, when you clicked the "Study" button this would call
startFlashcards and updateFlashcardsRecommendation() to progress the
flashcards recommendations. The combination of that button and the
recommendation progression on first render would inadvertently take 2
steps instead of 1. Now, we don't explicitly update the recommendation
on click of "Study" and instead rely on the useEffect calling
updateFlashcardsRecommendation() when flashcardsCourseLevel changes and
on first render.

Altogether, we progress the recommendation by 1 step when:

- you click "Study" and then navigate back to Progress page
- you click "Skip"
- you select a flashcards level