Fix matchSplitText tests
The spacePlacement value in test fixtures was wrong, so these tests
inadvertently passed even though they were written incorrectly. The
input argument for "expected" text is passed to the method without any
spacing. Fixing the spacePlacement values and removing the spaces in the
"expected" values used in tests makes the tests pass and behave
accurately.